+++ OPCSensor.h
@@ -45,8 +45,8 @@ class OPC																//Parent OPC class
 	String CSVHeader();													//Placeholders
 	String logUpdate();													
 	bool readData();
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill);
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);
+	void getData(float *dataPtr[], unsigned int arrayFill);
+	void getData(float *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);
 	void setReset(unsigned long resetTimer);							//Manually set the bad log reset timer
 };
 
@@ -68,8 +68,8 @@ class Plantower: public OPC
 	String CSVHeader();													//Overrides of OPC data functions
 	String logUpdate();
 	bool readData();
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill);			//Get data will pass the data into an array via a pointer
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);
+	void getData(float *dataPtr[], unsigned int arrayFill);				//Get data will pass the data into an array via a pointer
+	void getData(float *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);
 };
 
 class SPS: public OPC
@@ -125,8 +125,8 @@ class R1: public OPC {													//The R1 runs on SPI Communication
 	String CSVHeader();													//Overrrides the OPC data functions
 	String logUpdate();
 	bool readData();
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill);			//Get data will pass the data into an array via a pointer
-	void getData(uint16_t *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);													
+	void getData(float *dataPtr[], unsigned int arrayFill);				//Get data will pass the data into an array via a pointer
+	void getData(float *dataPtr[], unsigned int arrayFill, unsigned int arrayStart);													
 };
 
 #endif
